tty: serial: tegra: check function's return value
Platform_get_irq returns int. So, the check should
be (int)irq < 0 not irq < 0 which has no effect as irq
is unsigned.

Bug 1046331

Reviewed-on: http://git-master/r/159553

+++ drivers/tty/serial/tegra_hsuart.c
@@ -1573,7 +1573,7 @@ static int __init tegra_uart_probe(struct platform_device *pdev)
 	u->iotype = UPIO_MEM32;
 
 	u->irq = platform_get_irq(pdev, 0);
-	if (unlikely(u->irq < 0)) {
+	if (unlikely((int)(u->irq) < 0)) {
 		ret = -ENXIO;
 		goto fail;
 	}
